feat: move network selector inside transactions tab
